fix several memory bugs and heap overflows pointed out by asan replace customised cJSON with latest upstream to fix heap corruption issues fix a bunch of crashes in bot waypoint code, remove some RMG code. fixes #351 fix crash in glass break code fix cg_forceOwnSaber, cg_forceEnemySaber not updating clientinfo fix cg_remaps not working in all cases fix linefeeds not functioning in g_motd fix JPLua keydown event calling wrong listener. fixes #440 fix some JA+ special move prediction fix japp_duelStats not taking ties into account fix crash caused by inflicting admin command on user with manually granted admin permissions. fixes #442 fix Jedi Master gametype not functioning if players spawn with a saber. fixes #426 fix bogus userinfo value for siege desired team fix incorrect memset on non-POD type `level` by moving reserved entities elsewhere. fixes #439 fix UI caching master server addresses incorrectly revert change in SelectRandomFurthestSpawnPoint causing off-by-one error. fixes #441 add japp_noFXRunner add cg_deferPlayers 2 to load deferred when not moving add bot_fallbackNoWP to stop bots jumping randomly without a waypoint

+++ cgame/cg_media.cpp
@@ -930,7 +930,7 @@ static void CG_RegisterModels( void ) {
 		vector3 mins, maxs;
 		int j;
 
-		Com_sprintf( name, sizeof(name), "*%i", i );
+		Com_sprintf( name, sizeof(name), "*%i", (int)i );
 		cgs.inlineDrawModel[i] = trap->R_RegisterModel( name );
 		if ( !cgs.inlineDrawModel[i] ) {
 			breakPoint = i;
@@ -981,7 +981,7 @@ static void CG_RegisterModels( void ) {
 
 		char loadingStr[1024];
 		Com_sprintf( loadingStr, sizeof(loadingStr), "Loading BSP instance %i/%i: %s",
-			i, numBSPInstances, bspName+1
+			(int)i, (int)numBSPInstances, bspName+1
 		);
 		CG_LoadingString( loadingStr );
 
@@ -996,7 +996,7 @@ static void CG_RegisterModels( void ) {
 
 		for ( int sub = 1; sub < MAX_MODELS; sub++ ) {
 			char temp[MAX_QPATH];
-			Com_sprintf( temp, MAX_QPATH, "*%d-%d", i, sub );
+			Com_sprintf( temp, MAX_QPATH, "*%d-%d", (int)i, sub );
 			cgs.inlineDrawModel[breakPoint] = trap->R_RegisterModel( temp );
 			if ( !cgs.inlineDrawModel[breakPoint] ) {
 				break;
